python/requirements.txt
python-telegram-bot
configparser
pymssql